+++ common/src/main/java/smartin/miapi/modules/material/palette/MaterialPaletteFromTexture.java
@@ -12,9 +12,7 @@ import net.minecraft.client.texture.SpriteDimensions;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
 import net.minecraft.util.math.ColorHelper;
-import net.minecraft.util.math.Direction;
 import org.jetbrains.annotations.Nullable;
-import org.joml.Vector4i;
 import smartin.miapi.Miapi;
 import smartin.miapi.mixin.client.SpriteContentsAccessor;
 import smartin.miapi.modules.material.Material;
@@ -26,25 +24,12 @@ public class MaterialPaletteFromTexture extends SimpleMaterialPalette {
     Supplier<NativeImage> imageSupplier;
 
     @Environment(EnvType.CLIENT)
-    public static MaterialColorer forGeneratedMaterial(Material material, ItemStack mainIngredient) {
-        try {
-            return new MaterialPaletteFromTexture(material, () -> {
-                BakedModel itemModel = MinecraftClient.getInstance().getItemRenderer().getModel(mainIngredient, MinecraftClient.getInstance().world, null, 0);
-                SpriteContents contents = itemModel.getParticleSprite().getContents();
-                return ((SpriteContentsAccessor) contents).getImage();
-            });
-        } catch (Exception surpressed) {
-            try {
-                return new MaterialPaletteFromTexture(material, () -> {
-                    BakedModel itemModel = MinecraftClient.getInstance().getItemRenderer().getModel(mainIngredient, MinecraftClient.getInstance().world, null, 0);
-                    SpriteContents contents = itemModel.getQuads(null, Direction.NORTH, MinecraftClient.getInstance().textRenderer.random).get(0).getSprite().getContents();
-                    return ((SpriteContentsAccessor) contents).getImage();
-                });
-            } catch (Exception e) {
-                Miapi.LOGGER.warn("Error during palette creation", e);
-            }
-            return new EmptyMaterialPalette(material);
-        }
+    public static MaterialPaletteFromTexture forGeneratedMaterial(Material material, ItemStack mainIngredient) {
+        return new MaterialPaletteFromTexture(material, () -> {
+            BakedModel itemModel = MinecraftClient.getInstance().getItemRenderer().getModel(mainIngredient, MinecraftClient.getInstance().world, null, 0);
+            SpriteContents contents = itemModel.getParticleSprite().getContents();
+            return ((SpriteContentsAccessor) contents).getImage();
+        });
     }
 
     public MaterialPaletteFromTexture(Material material, Supplier<NativeImage> img) {
@@ -53,63 +38,44 @@ public class MaterialPaletteFromTexture extends SimpleMaterialPalette {
         this.setSpriteId(new Identifier(Miapi.MOD_ID, "miapi_materials/" + material.getKey()));
     }
 
-    public static List<Color> sortedColors(List<Color> colors) {
-        final Color[] prev = {null};
-        return colors.stream()
-                .sorted(Comparator.comparingDouble(Vector4i::length))
-                .sorted(Comparator.comparingDouble((a) -> {
-                    if (prev[0] == null) {
-                        prev[0] = a;
-                        return 0;
-                    }
-                    double distanceToPrevious = distanceBetween(a, prev[0]);
-                    prev[0] = a;
-                    return distanceToPrevious;
-                }))
+    public Map<Integer,Color> getColorPalette(){
+        List<Color> pixels = Arrays.stream(getPixelArray())
+                .mapToObj(Color::new)
+                .filter(color -> color.a() > 5)
+                .sorted(Comparator.comparingDouble((a) -> a.toFloatVecNoDiv().length()))
                 .toList();
-    }
+        Map<Integer, Color> finalColorMap = new LinkedHashMap<>();
+        List<Color> uniqueColors = pixels.stream().distinct().toList();
+        for (int i = 0; i < uniqueColors.size() - 1; i++) {
+            float scale = 256.0f / (float) pixels.size();
+            Color current = uniqueColors.get(i);
 
-    public static double distanceBetween(Color v1, Color v2) {
-        double dx = v1.x - v2.x;
-        double dy = v1.y - v2.y;
-        double dz = v1.z - v2.z;
-        double dw = v1.w - v2.w;
-        return Math.sqrt(dx * dx + dy * dy + dz * dz + dw * dw);
+            float distanceCurrent = Math.max(1, pixels.lastIndexOf(current) - pixels.indexOf(current));
+
+            int medianScaled = (int) (pixels.indexOf(current) + distanceCurrent / 2);
+            finalColorMap.put((int) ((medianScaled) * scale), current);
+
+            Color next = uniqueColors.get(i + 1);
+            float distanceNext = Math.max(1, pixels.lastIndexOf(next) - pixels.indexOf(next));
+            //float weight = Math.min(1, Math.max(0, Math.min(distanceCurrent / distanceNext, 1 - distanceNext / distanceCurrent)));
+            float weight = Math.min(1, Math.max(0, distanceCurrent / (distanceCurrent + distanceNext)));
+            Color weightedAverage = new Color(
+                    current.redAsFloat() * weight + next.redAsFloat() * (1 - weight),
+                    current.greenAsFloat() * weight + next.greenAsFloat() * (1 - weight),
+                    current.blueAsFloat() * weight + next.blueAsFloat() * (1 - weight),
+                    current.alphaAsFloat() * weight + next.alphaAsFloat() * (1 - weight));
+            float weightedPos = pixels.lastIndexOf(current) + (pixels.lastIndexOf(current) - pixels.indexOf(next)) * 0.5f;
+            finalColorMap.put((int) (weightedPos * scale), weightedAverage);
+        }
+        finalColorMap.putIfAbsent(0, Color.BLACK);
+        finalColorMap.putIfAbsent(255, uniqueColors.get(uniqueColors.size() - 1));
+        return finalColorMap;
     }
 
     @Environment(EnvType.CLIENT)
     public @Nullable SpriteContents generateSpriteContents(Identifier id) {
         try {
-            List<Color> pixels = Arrays.stream(getPixelArray())
-                    .mapToObj(Color::new)
-                    .filter(color -> color.a() > 5)
-                    .sorted(Comparator.comparingDouble((a) -> a.toFloatVecNoDiv().length()))
-                    .toList();
-            Map<Integer, Color> finalColorMap = new HashMap<>();
-            List<Color> uniqueColors = pixels.stream().distinct().toList();
-            for (int i = 0; i < uniqueColors.size() - 1; i++) {
-                float scale = 256.0f / (float) pixels.size();
-                Color current = uniqueColors.get(i);
-
-                float distanceCurrent = Math.max(1, pixels.lastIndexOf(current) - pixels.indexOf(current));
-
-                int medianScaled = (int) (pixels.indexOf(current) + distanceCurrent / 2);
-                finalColorMap.put((int) ((medianScaled) * scale), current);
-
-                Color next = uniqueColors.get(i + 1);
-                float distanceNext = Math.max(1, pixels.lastIndexOf(next) - pixels.indexOf(next));
-                //float weight = Math.min(1, Math.max(0, Math.min(distanceCurrent / distanceNext, 1 - distanceNext / distanceCurrent)));
-                float weight = Math.min(1, Math.max(0, distanceCurrent / (distanceCurrent + distanceNext)));
-                Color weightedAverage = new Color(
-                        current.redAsFloat() * weight + next.redAsFloat() * (1 - weight),
-                        current.greenAsFloat() * weight + next.greenAsFloat() * (1 - weight),
-                        current.blueAsFloat() * weight + next.blueAsFloat() * (1 - weight),
-                        current.alphaAsFloat() * weight + next.alphaAsFloat() * (1 - weight));
-                float weightedPos = pixels.lastIndexOf(current) + (pixels.lastIndexOf(current) - pixels.indexOf(next)) * 0.5f;
-                finalColorMap.put((int) (weightedPos * scale), weightedAverage);
-            }
-            finalColorMap.putIfAbsent(0, Color.BLACK);
-            finalColorMap.putIfAbsent(255, uniqueColors.get(uniqueColors.size() - 1));
+            Map<Integer, Color> finalColorMap = getColorPalette();
             NativeImage image = new NativeImage(256, 1, false);
             PaletteCreators.PixelPlacer placer = (color, x, y) -> image.setColor(x, y, color.abgr());
             PaletteCreators.FillerFunction filler = PaletteCreators.fillers.getOrDefault("interpolation", PaletteCreators.interpolateFiller);